fixed: camelCase & Schemas Refactor
