+++ docs/paths/Clinic/clinicsLogin.js
