+++ docs/paths/Clinic/clinicsResend.js
