+++ docs/paths/Clinic/clinicsVerify.js
